+++ src/di/tests/Stub/ProxyTraitObject.php
@@ -17,6 +17,16 @@ class ProxyTraitObject
 {
     use ProxyTrait;
 
+    /**
+     * @var string
+     */
+    public $name;
+
+    public function __construct(string $name = 'Hyperf')
+    {
+        $this->name = $name;
+    }
+
     public function get(?int $id, string $str = '')
     {
         return $this->__getParamsMap(static::class, 'get', func_get_args());
@@ -34,10 +44,22 @@ class ProxyTraitObject
 
     public function incr()
     {
-        $__function__ = __FUNCTION__;
-        $__method__ = __METHOD__;
-        return self::__proxyCall(ProxyTraitObject::class, __FUNCTION__, self::__getParamsMap(ProxyTraitObject::class, __FUNCTION__, func_get_args()), function () use ($__function__, $__method__) {
+        return self::__proxyCall(ProxyTraitObject::class, __FUNCTION__, self::__getParamsMap(ProxyTraitObject::class, __FUNCTION__, func_get_args()), function () {
             return 1;
+        });
+    }
+
+    public function getName()
+    {
+        return self::__proxyCall(ProxyTraitObject::class, __FUNCTION__, self::__getParamsMap(ProxyTraitObject::class, __FUNCTION__, func_get_args()), function () {
+            return 'HyperfCloud';
+        });
+    }
+
+    public function getName2()
+    {
+        return self::__proxyCall(ProxyTraitObject::class, __FUNCTION__, self::__getParamsMap(ProxyTraitObject::class, __FUNCTION__, func_get_args()), function () {
+            return 'HyperfCloud';
         });
     }
 }