+++ src/guzzle/src/CoroutineHandler.php
@@ -190,8 +190,9 @@ class CoroutineHandler
         }
 
         $body = $client->body;
-        if (isset($options['sink']) && is_string($options['sink'])) {
-            $body = $this->createSink($body, $options['sink']);
+        $sink = $options['sink'] ?? null;
+        if (isset($sink) && (is_string($sink) || is_resource($sink))) {
+            $body = $this->createSink($body, $sink);
         }
 
         $response = new Psr7\Response(
@@ -220,16 +221,16 @@ class CoroutineHandler
         return Psr7\stream_for($body);
     }
 
-    protected function createSink(string $body, string $sink)
+    /**
+     * @param resource|string $stream
+     */
+    protected function createSink(string $body, $stream)
     {
-        if (! empty($options['stream'])) {
-            return $body;
+        if (is_string($stream)) {
+            $stream = fopen($stream, 'w+');
         }
-
-        $stream = fopen($sink, 'w+');
         if ($body !== '') {
             fwrite($stream, $body);
-            fseek($stream, 0);
         }
 
         return $stream;
@@ -244,18 +245,21 @@ class CoroutineHandler
             'errCode' => $errCode,
         ];
 
-        if ($statusCode === SWOOLE_HTTP_CLIENT_ESTATUS_CONNECT_FAILED) {
-            return new ConnectException(sprintf('Connection failed, errCode=%s', $errCode), $request, null, $ctx);
+        if ($statusCode > 0) {
+            return true;
         }
 
-        if ($statusCode === SWOOLE_HTTP_CLIENT_ESTATUS_REQUEST_TIMEOUT) {
-            return new RequestException(sprintf('Request timed out, errCode=%s', $errCode), $request, null, null, $ctx);
+        switch ($statusCode) {
+            case SWOOLE_HTTP_CLIENT_ESTATUS_CONNECT_FAILED:
+                return new ConnectException(sprintf('Connection failed, errCode=%s', $errCode), $request, null, $ctx);
+            case SWOOLE_HTTP_CLIENT_ESTATUS_REQUEST_TIMEOUT:
+                return new RequestException(sprintf('Request timed out, errCode=%s', $errCode), $request, null, null, $ctx);
+            case SWOOLE_HTTP_CLIENT_ESTATUS_SERVER_RESET:
+                return new RequestException('Server reset', $request, null, null, $ctx);
+            case -4:
+                return new RequestException(sprintf('Send failed, errCode=%s', $errCode), $request, null, null, $ctx);
+            default:
+                return new RequestException(sprintf('Unknown reason, errCode=%s', $errCode), $request, null, null, $ctx);
         }
-
-        if ($statusCode === SWOOLE_HTTP_CLIENT_ESTATUS_SERVER_RESET) {
-            return new RequestException('Server reset', $request, null, null, $ctx);
-        }
-
-        return true;
     }
 }