+++ src/view-engine/src/Compiler/ComponentTagCompiler.php
@@ -51,24 +51,35 @@ class ComponentTagCompiler
     protected $boundAttributes = [];
 
     /**
+     * @var array
+     */
+    protected $autoloadClasses = [];
+
+    /**
+     * @var array
+     */
+    protected $autoloadComponents = [];
+
+    /**
      * Create new component tag compiler.
      *
      * @param ?BladeCompiler $blade
+     * @param mixed $autoload
      */
-    public function __construct(array $aliases = [], array $namespaces = [], ?BladeCompiler $blade = null)
+    public function __construct(array $aliases = [], array $namespaces = [], ?BladeCompiler $blade = null, $autoload = [])
     {
         $this->aliases = $aliases;
         $this->namespaces = $namespaces;
+        $this->autoloadClasses = $autoload['classes'] ?? [null];
+        $this->autoloadComponents = $autoload['components'] ?? [null];
 
         $this->blade = $blade ?: new BladeCompiler(new Filesystem(), sys_get_temp_dir());
     }
 
     /**
      * Compile the component and slot tags within the given string.
-     *
-     * @return string
      */
-    public function compile(string $value)
+    public function compile(string $value): string
     {
         $value = $this->compileSlots($value);
 
@@ -79,9 +90,8 @@ class ComponentTagCompiler
      * Compile the tags within the given string.
      *
      * @throws InvalidArgumentException
-     * @return string
      */
-    public function compileTags(string $value)
+    public function compileTags(string $value): string
     {
         $value = $this->compileSelfClosingTags($value);
         $value = $this->compileOpeningTags($value);
@@ -92,9 +102,8 @@ class ComponentTagCompiler
      * Get the component class for a given component alias.
      *
      * @throws InvalidArgumentException
-     * @return string
      */
-    public function componentClass(string $component)
+    public function componentClass(string $component): string
     {
         $viewFactory = Blade::container()->get(FactoryInterface::class);
 
@@ -116,11 +125,7 @@ class ComponentTagCompiler
             return $class;
         }
 
-        if (class_exists($class = $this->guessClassName($component))) {
-            return $class;
-        }
-
-        if ($viewFactory->exists($view = $this->guessViewName($component))) {
+        if ($view = $this->guessComponentFromAutoload($viewFactory, $component)) {
             return $view;
         }
 
@@ -130,43 +135,65 @@ class ComponentTagCompiler
     }
 
     /**
-     * Find the class for the given component using the registered namespaces.
+     * Guess the view or class name for the given component.
      *
-     * @return null|string|void
+     * @return string
      */
-    public function findClassByComponent(string $component)
+    public function guessComponentFromAutoload(FactoryInterface $viewFactory, string $component): ?string
+    {
+        foreach ($this->autoloadClasses ?: [null] as $prefix) {
+            if (class_exists($view = $this->guessClassName($component, (string) $prefix))) {
+                return $view;
+            }
+        }
+
+        foreach ($this->autoloadComponents ?: [null] as $prefix) {
+            if ($viewFactory->exists($view = $this->guessViewName($component, (string) $prefix))) {
+                return $view;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Find the class for the given component using the registered namespaces.
+     */
+    public function findClassByComponent(string $component): ?string
     {
         $segments = explode('::', $component);
 
         $prefix = $segments[0];
 
         if (! isset($this->namespaces[$prefix]) || ! isset($segments[1])) {
-            return;
+            return null;
         }
 
         if (class_exists($class = $this->namespaces[$prefix] . '\\' . $this->formatClassName($segments[1]))) {
             return $class;
         }
+
+        return null;
     }
 
     /**
      * Guess the class name for the given component.
-     *
-     * @return string
      */
-    public function guessClassName(string $component)
+    public function guessClassName(string $component, string $prefix = ''): string
     {
         $class = $this->formatClassName($component);
 
-        return 'App\\View\\Components\\' . $class;
+        if (! $prefix) {
+            $prefix = 'App\\View\\Component\\';
+        }
+
+        return  rtrim($prefix, '\\') . '\\' . $class;
     }
 
     /**
      * Format the class name for the given component.
-     *
-     * @return string
      */
-    public function formatClassName(string $component)
+    public function formatClassName(string $component): string
     {
         $componentPieces = array_map(function ($componentPiece) {
             return ucfirst(Str::camel($componentPiece));
@@ -177,13 +204,10 @@ class ComponentTagCompiler
 
     /**
      * Guess the view name for the given component.
-     *
-     * @param string $name
-     * @return string
      */
-    public function guessViewName($name)
+    public function guessViewName(string $name, string $prefix = ''): string
     {
-        $prefix = 'components.';
+        $prefix = $prefix ?: 'components.';
 
         $delimiter = FinderInterface::HINT_PATH_DELIMITER;
 
@@ -196,11 +220,8 @@ class ComponentTagCompiler
 
     /**
      * Partition the data and extra attributes from the given array of attributes.
-     *
-     * @param string $class
-     * @return array
      */
-    public function partitionDataAndAttributes($class, array $attributes)
+    public function partitionDataAndAttributes(string $class, array $attributes): array
     {
         // If the class doesn't exists, we'll assume it's a class-less component and
         // return all of the attributes as both data and attributes since we have
@@ -222,10 +243,8 @@ class ComponentTagCompiler
 
     /**
      * Compile the slot tags within the given string.
-     *
-     * @return string
      */
-    public function compileSlots(string $value)
+    public function compileSlots(string $value): string
     {
         $value = preg_replace_callback('/<\s*x[\-\:]slot\s+(:?)name=(?<name>(\"[^\"]+\"|\\\'[^\\\']+\\\'|[^\s>]+))\s*>/', function ($matches) {
             $name = $this->stripQuotes($matches['name']);
@@ -242,10 +261,8 @@ class ComponentTagCompiler
 
     /**
      * Strip any quotes from the given string.
-     *
-     * @return string
      */
-    public function stripQuotes(string $value)
+    public function stripQuotes(string $value): string
     {
         return Str::startsWith($value, ['"', '\''])
             ? substr($value, 1, -1)
@@ -256,9 +273,8 @@ class ComponentTagCompiler
      * Compile the opening tags within the given string.
      *
      * @throws InvalidArgumentException
-     * @return string
      */
-    protected function compileOpeningTags(string $value)
+    protected function compileOpeningTags(string $value): string
     {
         $pattern = "/
             <
@@ -306,9 +322,8 @@ class ComponentTagCompiler
      * Compile the self-closing tags within the given string.
      *
      * @throws InvalidArgumentException
-     * @return string
      */
-    protected function compileSelfClosingTags(string $value)
+    protected function compileSelfClosingTags(string $value): string
     {
         $pattern = "/
             <
@@ -356,9 +371,8 @@ class ComponentTagCompiler
      * Compile the Blade component string for the given component and attributes.
      *
      * @throws InvalidArgumentException
-     * @return string
      */
-    protected function componentString(string $component, array $attributes)
+    protected function componentString(string $component, array $attributes): string
     {
         $class = $this->componentClass($component);
 
@@ -388,20 +402,16 @@ class ComponentTagCompiler
 
     /**
      * Compile the closing tags within the given string.
-     *
-     * @return string
      */
-    protected function compileClosingTags(string $value)
+    protected function compileClosingTags(string $value): string
     {
         return preg_replace('/<\\/\\s*x[-\\:][\\w\\-\\:\\.]*\\s*>/', ' @endcomponentClass ', $value);
     }
 
     /**
      * Get an array of attributes from the given attribute string.
-     *
-     * @return array
      */
-    protected function getAttributesFromAttributeString(string $attributeString)
+    protected function getAttributesFromAttributeString(string $attributeString): array
     {
         $attributeString = $this->parseAttributeBag($attributeString);
 
@@ -453,10 +463,8 @@ class ComponentTagCompiler
 
     /**
      * Parse the attribute bag in a given attribute string into it's fully-qualified syntax.
-     *
-     * @return string
      */
-    protected function parseAttributeBag(string $attributeString)
+    protected function parseAttributeBag(string $attributeString): string
     {
         $pattern = '/
             (?:^|\\s+)                                        # start of the string or whitespace between attributes
@@ -468,10 +476,8 @@ class ComponentTagCompiler
 
     /**
      * Parse the "bind" attributes in a given attribute string into their fully-qualified syntax.
-     *
-     * @return string
      */
-    protected function parseBindAttributes(string $attributeString)
+    protected function parseBindAttributes(string $attributeString): string
     {
         $pattern = '/
             (?:^|\\s+)     # start of the string or whitespace between attributes
@@ -487,10 +493,8 @@ class ComponentTagCompiler
      * Compile any Blade echo statements that are present in the attribute string.
      *
      * These echo statements need to be converted to string concatenation statements.
-     *
-     * @return string
      */
-    protected function compileAttributeEchos(string $attributeString)
+    protected function compileAttributeEchos(string $attributeString): string
     {
         $value = $this->blade->compileEchos($attributeString);
 
@@ -502,10 +506,8 @@ class ComponentTagCompiler
 
     /**
      * Escape the single quotes in the given string that are outside of PHP blocks.
-     *
-     * @return string
      */
-    protected function escapeSingleQuotesOutsideOfPhpBlocks(string $value)
+    protected function escapeSingleQuotesOutsideOfPhpBlocks(string $value): string
     {
         return collect(token_get_all($value))->map(function ($token) {
             if (! is_array($token)) {
@@ -520,11 +522,8 @@ class ComponentTagCompiler
 
     /**
      * Convert an array of attributes to a string.
-     *
-     * @param bool $escapeBound
-     * @return string
      */
-    protected function attributesToString(array $attributes, $escapeBound = true)
+    protected function attributesToString(array $attributes, bool $escapeBound = true): string
     {
         return collect($attributes)
             ->map(function (string $value, string $attribute) use ($escapeBound) {